feat: swap task master with quersi
- remove task master
- add quersi
- remove unused files related to task master
