+++ mobile-app/lib/services/wallet_creation_service.dart
@@ -1,20 +1,16 @@
 import 'dart:async';
 
 import 'package:quantus_sdk/quantus_sdk.dart';
-import 'package:resonance_network_wallet/services/referral_service.dart';
 
 class WalletCreationService {
   final SettingsService _settings;
   final AccountsService _accounts;
-  final ReferralService _referral;
 
   WalletCreationService({
     SettingsService? settingsService,
     AccountsService? accountsService,
-    ReferralService? referralService,
   }) : _settings = settingsService ?? SettingsService(),
-       _accounts = accountsService ?? AccountsService(),
-       _referral = referralService ?? ReferralService();
+       _accounts = accountsService ?? AccountsService();
 
   /// Saves [mnemonic] for [walletIndex], adds the root account when missing,
   /// and runs referral registration for brand-new roots.
@@ -35,7 +31,6 @@ class WalletCreationService {
       _settings.setWalletOrigin(walletIndex, WalletOrigin.created);
       final account = Account(walletIndex: walletIndex, index: 0, name: name, accountId: accountId);
       await _accounts.addAccount(account);
-      unawaited(_referral.submitAddressToBackend());
       return account;
     }
